+++ nprogress.mjs
@@ -1,0 +1,6 @@
+import * as moduleNamespace from './nprogress.js';
+
+const NProgress = globalThis.NProgress || moduleNamespace.default || moduleNamespace;
+
+export { NProgress };
+export default NProgress;